requirements.txt
fastapi
python-dotenv
SQLAlchemy
asyncpg
uvicorn
pydantic
bcrypt